Add "Silenced" state for Monitoring Alerts
Introduce a "Silenced" Alert state for Alerts that are silenced by at
least one active Silence. Allow filtering the Alerts list by this
new status.

Add Silenced By list to the Alert details page that shows which
Silence(s) are silencing the Alert.

Add Silenced Alerts list to the Silence details page that shows which
Alerts the Silence silences.

Combine components/alert.tsx and components/silence.tsx into a single
file since the Silence details page now includes Alerts elements and the
Alert details page includes Silence elements.
