+++ frontend/public/components/app.jsx
@@ -12,13 +12,12 @@ import { ALL_NAMESPACES_KEY } from '../const';
 import { connectToFlags, featureActions, flagPending, FLAGS } from '../features';
 import { detectMonitoringURLs } from '../monitoring';
 import { analyticsSvc } from '../module/analytics';
-import { AlertsPage, AlertsDetailsPage, AlertRulesDetailsPage } from './alert';
+import { AlertsPage, AlertsDetailsPage, AlertRulesDetailsPage, CreateSilence, EditSilence, SilencesPage, SilencesDetailsPage } from './monitoring';
 import { GlobalNotifications } from './global-notifications';
 import { Masthead } from './masthead';
 import { NamespaceBar } from './namespace';
 import { Nav } from './nav';
 import { SearchPage } from './search';
-import { CreateSilence, EditSilence, SilencesPage, SilencesDetailsPage } from './silence';
 import { ResourceDetailsPage, ResourceListPage } from './resource-list';
 import { history, AsyncComponent, Loading } from './utils';
 import { namespacedPrefixes } from './utils/link';
